requirements.txt
asyncio
random
discord.py
websockets
chardet
aiohttp